requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow==2.10.0
opencv-python==4.8.0.74
numpy==1.24.3
h5py==3.8.0
Pillow==9.5.0
gunicorn==20.1.0
protobuf==3.20.3
astunparse==1.6.3
flatbuffers==2.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.48.2
libclang==13.0.0
opt-einsum==3.3.0
termcolor==2.3.0
typing_extensions==4.3.0
wrapt==1.14.1
absl-py==1.4.0
keras==2.10.0
tensorboard==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0